Bunch of changes
Uhh so completely switched how I call the steps counter. Plus this fixes the timing of the fall. Also added layers to the different materials so now the step sound changes. Only issue is that I somehow broke the jump function of the player soooooo. help. Additionally, need to figure out how to mute the step sounds while in the air